restructure and fix merge-conflicts
