+++ services/apps/script_executor_worker/src/workflows/fixDeletedMemberOrgAffilations.ts
@@ -1,5 +1,6 @@
 import { continueAsNew, proxyActivities } from '@temporalio/workflow'
 
+import * as commonActivities from '../activities/common'
 import * as activities from '../activities/fix-deleted-member-org-affilations'
 import * as syncActivities from '../activities/sync/member'
 import { IFixDeletedMemberOrgAffilationsArgs } from '../types'
@@ -14,6 +15,11 @@ const syncActivity = proxyActivities<typeof syncActivities>({
   retry: { maximumAttempts: 3, backoffCoefficient: 3 },
 })
 
+const commonActivity = proxyActivities<typeof commonActivities>({
+  startToCloseTimeout: '30 minutes',
+  retry: { maximumAttempts: 3, backoffCoefficient: 3 },
+})
+
 export async function fixDeletedMemberOrgAffilations(
   args: IFixDeletedMemberOrgAffilationsArgs,
 ): Promise<void> {
@@ -23,7 +29,7 @@ export async function fixDeletedMemberOrgAffilations(
     `Fixing deleted member org affiliations with params: testRun=${args.testRun}, batchSize=${args.batchSize}`,
   )
 
-  // 1. Find affected memberId and orgId
+  // Find affected memberId and orgId
   const affectedMembers = await activity.getProcessedMemberOrgAffiliations(BATCH_SIZE)
 
   if (affectedMembers.length === 0) {
@@ -36,16 +42,6 @@ export async function fixDeletedMemberOrgAffilations(
     const chunk = affectedMembers.slice(i, i + CHUNK_SIZE)
     await Promise.all(
       chunk.map(async ({ memberId, organizationId }) => {
-        // 2. Check if they have activity in questDb
-        // const hasActivity = await activity.hasActivityInQuestDb(memberId, organizationId)
-
-        // 2.1 If no activities found, we need to get and insert them
-        // if (!hasActivity) {
-        //   console.log(`Copying activities for member ${memberId} and org ${organizationId}`)
-
-        //   await activity.copyActivitiesFromPgToQuestDb(memberId, organizationId)
-        // }
-
         // Calculate affiliation
         await activity.calculateMemberAffiliations(memberId)
 
@@ -54,7 +50,7 @@ export async function fixDeletedMemberOrgAffilations(
 
         // Add orgId to redisCache
         // It will be picked up by the spawnOrganizationAggregatesComputation workflow
-        await activity.addOrgIdToRedisCache(organizationId)
+        await commonActivity.queueOrganizationForAggComputation(organizationId)
 
         await activity.deleteProcessedMemberOrgAffiliations(memberId, organizationId)
       }),
